07/11/2023

+++ controller/user.controller.js
@@ -1,16 +1,20 @@
 const { ResponseTemplate } = require('../helper/template.helper')
-const { PrismaClient} = require('@prisma/client')
+const { ComparePassword, HashPassword } = require('../helper/hash_pass_helper')
+const { PrismaClient } = require('@prisma/client')
 
 const prisma = new PrismaClient()
+var jwt = require('jsonwebtoken')
 
 async function Insert(req, res) {
 
     const { name, email, password, identity_type, identity_number, address } = req.body
 
+    const hashPass = await HashPassword(password)
+
     const payload = {
         name,
         email,
-        password,
+        password: hashPass,
         profile: {
             create: {
                 identity_type,
@@ -20,16 +24,46 @@ async function Insert(req, res) {
         }
     }
 
+    console.log(payload)
+
+    const emailUser = await prisma.user.findUnique({
+        where: {email: payload.email},
+    });
+
+    if (emailUser) {
+        let resp = ResponseTemplate(null, 'Email already exist', null, 404)
+        res.json(resp)
+        return
+    }
+
     try {
         
-        const user = await prisma.user.create({
+        await prisma.user.create({
             data: payload,
             include: {
                 profile: true
             }
         });
 
-        let resp = ResponseTemplate(user, 'success', null, 200)
+        const userView = await prisma.user.findUnique({
+            where: {
+                email: payload.email
+            },
+            select: {
+                id: true,
+                name: true,
+                email: true,
+                profile: {
+                    select: {
+                        identity_type: true,
+                        identity_number: true,
+                        address: true
+                    }
+                }
+            },
+        });
+
+        let resp = ResponseTemplate(userView, 'success', null, 200)
         res.json(resp);
         return
 
@@ -41,29 +75,59 @@ async function Insert(req, res) {
     }
 }
 
+async function Login(req, res) {
+
+    try {
+        const { email, password } = req.body
+
+        const checkUser = await prisma.user.findUnique({
+            where: {
+                email: email
+            }
+        })
+
+        if (checkUser === null) {
+            let resp = ResponseTemplate(null, 'email is not found or incorrect', null, 400)
+            res.json(resp)
+            return
+        }
+
+        const checkPassword = await ComparePassword(password, checkUser.password)
+
+        if (!checkPassword) {
+            let resp = ResponseTemplate(null, 'password is not correct', null, 400)
+            res.json(resp)
+            return
+        }
+
+        const token = jwt.sign({
+            email: checkUser.email,
+            user_id: checkUser.id
+        }, process.env.SECRET_KEY);
+
+        let resp = ResponseTemplate(token, 'success', null, 200)
+        res.json(resp)
+        return
+
+    } catch (error) {
+        let resp = ResponseTemplate(null, 'internal server error', error, 500)
+        res.json(resp)
+        return
+    }
+}
+
 async function Get(req, res) {
 
-    const { name, email, password } = req.query
+    const { name, email, password, page = 1, limit = 10 } = req.query
 
     const payload = {}
 
-    if (name) {
-        payload.name = name
-    }
-
-    if (email) {
-        payload.email = email
-    }
-
-    if (password) {
-        payload.password = password
-    }
+    if (name) payload.name = name
+    if (email) payload.email = email
+    if (password) payload.password = password
 
     try {
-        // let page =1
-        // let limit = 10
-        let { page = 1, limit = 10 } = req.query // menghasilkan string
-        let skip = ( page - 1 ) * limit
+        const skip = ( page - 1 ) * limit
 
         //informasi total data keseluruhan 
         const resultCount = await prisma.user.count() // integer jumlah total data user
@@ -90,15 +154,18 @@ async function Get(req, res) {
                 }
             },
         });
-
+        
         const pagination = {
             current_page: page - 0, // ini - 0 merubah menjadi integer
             total_page : totalPage,
             total_data: resultCount,
             data: users
         }
-
-        if (users === null) {
+        const cekUser = (objectName) => {
+            return Object.keys(objectName).length === 0
+        }
+        
+        if (cekUser(users) === true) {
             let resp = ResponseTemplate(null, 'data not found', null, 404)
             res.json(resp)
             return
@@ -112,8 +179,6 @@ async function Get(req, res) {
         let resp = ResponseTemplate(null, 'internal server error', error, 500)
         res.json(resp)
         return
-
-
     }
 }
 
@@ -154,43 +219,50 @@ async function GetByPK(req, res) {
         let resp = ResponseTemplate(null, 'internal server error', error, 500)
         res.json(resp)
         return
-
-
     }
 }
 
 async function Update(req, res) {
 
-    const { name, email, password } = req.body
+    const { name, email, password, identity_type, identity_number, address } = req.body
     const { id } = req.params
 
     const payload = {}
+    const update = {}
+    const profile = {update}
 
-    if (!name && !email && !password) {
+    if (!name && !email && !password && !identity_type && !identity_number && !address) {
         let resp = ResponseTemplate(null, 'bad request', null, 400)
         res.json(resp)
         return
     }
 
-    if (name) {
-        payload.name = name
-    }
-
-    if (email) {
-        payload.email = email
-    }
-
-    if (password) {
-        payload.address = password
-    }
-
+    if (name) payload.name = name
+    if (email) payload.email = email
+    if (password) payload.password = password
+    if (identity_type || identity_number || address) payload.profile = profile
+    if (identity_type) update.identity_type = identity_type
+    if (identity_number) update.identity_number = identity_number
+    if (address) update.address = address
 
     try {
         const users = await prisma.user.update({
             where: {
                 id: Number(id)
             },
-            data: payload
+            data: payload,
+            select: {
+                id: true,
+                name: true,
+                email: true,
+                profile: {
+                    select: {
+                        identity_type: true,
+                        identity_number: true,
+                        address: true
+                    }
+                }
+            }
         })
 
         let resp = ResponseTemplate(users, 'success', null, 200)
@@ -201,8 +273,6 @@ async function Update(req, res) {
         let resp = ResponseTemplate(null, 'internal server error', error, 500)
         res.json(resp)
         return
-
-
     }
 }
 
@@ -211,6 +281,19 @@ async function Delete(req, res) {
     const { id } = req.params
 
     try {
+
+        const checkUser = await prisma.user.findFirst({
+            where: {
+                id: Number(id)
+            }
+        })
+
+        if (checkUser === null) {
+            let resp = ResponseTemplate(null, 'data not found', null, 404)
+            res.json(resp)
+            return
+        }
+
         await prisma.profile.delete({
             where: {
                 user_id: Number(id)
@@ -237,6 +320,7 @@ async function Delete(req, res) {
 
 module.exports = {
     Insert,
+    Login,
     Get,
     GetByPK,
     Update,
